organizing menu and folder structure
