+++ src/utils/utils.h
@@ -1,0 +1,3 @@
+float distanceToEnemy(float pX, float pY, float eX, float eY);
+void renderMenu(Vector2 *mouse, int screenWidth, Rectangle menuOptions[], int *menuMode);
+void enemyDinamic(Rectangle player, Rectangle *enemy, int speed, float *playerLifeBar);
